KT - fixed table file
