+++ frontend/src/main/components/Hotels/HotelTable.js
@@ -1,40 +1,36 @@
 import React from "react";
 import OurTable, { ButtonColumn } from "main/components/OurTable";
+import { useBackendMutation } from "main/utils/useBackend";
+import { cellToAxiosParamsDelete, onDeleteSuccess } from "main/utils/hotelUtils"
 import { useNavigate } from "react-router-dom";
-import { hotelUtils } from "main/utils/hotelUtils";
+import { hasRole } from "main/utils/currentUser";
 
-const showCell = (cell) => JSON.stringify(cell.row.values);
-
-
-const defaultDeleteCallback = async (cell) => {
-    console.log(`deleteCallback: ${showCell(cell)})`);
-    hotelUtils.del(cell.row.values.id);
-}
-
-export default function HotelTable({
-    hotels,
-    deleteCallback = defaultDeleteCallback,
-    showButtons = true,
-    testIdPrefix = "HotelTable" }) {
+export default function HotelTable({ hotels, currentUser }) {
 
     const navigate = useNavigate();
- 
+
     const editCallback = (cell) => {
-        console.log(`editCallback: ${showCell(cell)})`);
         navigate(`/hotels/edit/${cell.row.values.id}`)
     }
 
-    const detailsCallback = (cell) => {
-        console.log(`detailsCallback: ${showCell(cell)})`);
-        navigate(`/hotels/details/${cell.row.values.id}`)
-    }
+    // Stryker disable all : hard to test for query caching
+
+    const deleteMutation = useBackendMutation(
+        cellToAxiosParamsDelete,
+        { onSuccess: onDeleteSuccess },
+        ["/api/hotels/all"]
+    );
+    // Stryker enable all 
+
+    // Stryker disable next-line all : TODO try to make a good test for this
+    const deleteCallback = async (cell) => { deleteMutation.mutate(cell); }
+
 
     const columns = [
         {
             Header: 'id',
             accessor: 'id', // accessor is the "key" in the data
         },
-
         {
             Header: 'Name',
             accessor: 'name',
@@ -43,27 +39,24 @@ export default function HotelTable({
             Header: 'Address',
             accessor: 'address',
         },
-
         {
             Header: 'Description',
             accessor: 'description',
         }
     ];
 
-    const buttonColumns = [
-        ...columns,
-        ButtonColumn("Details", "primary", detailsCallback, testIdPrefix),
-        ButtonColumn("Edit", "primary", editCallback, testIdPrefix),
-        ButtonColumn("Delete", "danger", deleteCallback, testIdPrefix),
-    ]
+    if (hasRole(currentUser, "ROLE_ADMIN")) {
+        columns.push(ButtonColumn("Edit", "primary", editCallback, "HotelTable"));
+        columns.push(ButtonColumn("Delete", "danger", deleteCallback, "HotelTable"));
+    } 
 
-    const columnsToDisplay = showButtons ? buttonColumns : columns;
+    // Stryker disable next-line ArrayDeclaration : [columns] is a performance optimization
+    const memoizedColumns = React.useMemo(() => columns, [columns]);
+    const memoizedHotels = React.useMemo(() => hotels, [hotels]);
 
     return <OurTable
-        data={hotels}
-        columns={columnsToDisplay}
-        testid={testIdPrefix}
+        data={memoizedHotels}
+        columns={memoizedColumns}
+        testid={"HotelTable"}
     />;
 };
-
-export { showCell };
