+++ frontend/src/main/components/Restaurants/RestaurantTable.js
@@ -3,25 +3,31 @@ import OurTable, { ButtonColumn } from "main/components/OurTable";
 import { useNavigate } from "react-router-dom";
 import { restaurantUtils } from "main/utils/restaurantUtils";
 
+const showCell = (cell) => JSON.stringify(cell.row.values);
 
-export default function RestaurantTable({ restaurants, currentUser }) {
+
+const defaultDeleteCallback = async (cell) => {
+    console.log(`deleteCallback: ${showCell(cell)})`);
+    restaurantUtils.del(cell.row.values.id);
+}
+
+export default function RestaurantTable({
+    restaurants,
+    deleteCallback = defaultDeleteCallback,
+    showButtons = true,
+    testIdPrefix = "RestaurantTable" }) {
+
     const navigate = useNavigate();
-
+ 
     const editCallback = (cell) => {
-        navigate(`/hotels/edit/${cell.row.values.id}`)
+        console.log(`editCallback: ${showCell(cell)})`);
+        navigate(`/restaurants/edit/${cell.row.values.id}`)
     }
 
-    // Stryker disable all : hard to test for query caching
-
-    const deleteMutation = useBackendMutation(
-        cellToAxiosParamsDelete,
-        { onSuccess: onDeleteSuccess },
-        ["/api/hotels/all"]
-    );
-    // Stryker enable all 
-
-    // Stryker disable next-line all : TODO try to make a good test for this
-    const deleteCallback = async (cell) => { deleteMutation.mutate(cell); }
+    const detailsCallback = (cell) => {
+        console.log(`detailsCallback: ${showCell(cell)})`);
+        navigate(`/restaurants/details/${cell.row.values.id}`)
+    }
 
     const columns = [
         {
@@ -36,25 +42,23 @@ export default function RestaurantTable({ restaurants, currentUser }) {
         {
             Header: 'Description',
             accessor: 'description',
-        },
-        {
-            Header: 'Address',
-            accessor: 'address',
         }
     ];
 
-    if (hasRole(currentUser, "ROLE_ADMIN")) {
-        columns.push(ButtonColumn("Edit", "primary", editCallback, "HotelTable"));
-        columns.push(ButtonColumn("Delete", "danger", deleteCallback, "HotelTable"));
-    } 
+    const buttonColumns = [
+        ...columns,
+        ButtonColumn("Details", "primary", detailsCallback, testIdPrefix),
+        ButtonColumn("Edit", "primary", editCallback, testIdPrefix),
+        ButtonColumn("Delete", "danger", deleteCallback, testIdPrefix),
+    ]
 
-    // Stryker disable next-line ArrayDeclaration : [columns] is a performance optimization
-    const memoizedColumns = React.useMemo(() => columns, [columns]);
-    const memoizedHotels = React.useMemo(() => hotels, [hotels]);
+    const columnsToDisplay = showButtons ? buttonColumns : columns;
 
     return <OurTable
-        data={memoizedHotels}
-        columns={memoizedColumns}
-        testid={"HotelTable"}
+        data={restaurants}
+        columns={columnsToDisplay}
+        testid={testIdPrefix}
     />;
 };
+
+export { showCell };
